Downloading hardcoded scheme form from database  to UI
